shard out the features generation code into its own file in lib so it can be shared with the explorer
